category list 통합 및 수입/지출 통합
