+++ app/AddIncome.tsx
@@ -1,115 +1,39 @@
-import React, { useState } from "react";
-import { KeyboardAvoidingView, ScrollView, StyleSheet, Text, TextInput, View } from "react-native";
-import { router } from 'expo-router';
-import AsyncStorage from '@react-native-async-storage/async-storage';
-import AddIncomeList from "../components/AddIncome/list";
-import Button from "../components/Button";
-import MemoModal from "../components/memoModal";
-import { useDispatch } from "react-redux";
-import { AppDispatch, RootState } from "../redux/store";
-import { userInputSlice } from "../redux/slice/userInput";
+import React from "react";
+import AddHistory from "../components/History/AddHistory";
+import { FontAwesome5, MaterialIcons, AntDesign } from '@expo/vector-icons';
 import { useSelector } from "react-redux";
+import { RootState } from "../redux/store";
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignContent: "center",
-  },
-  priceInputContainer: {
-    flexDirection: "row",
-    width: "100%",
-    justifyContent: "center",
-    flex: 0.15,
-    alignItems: "center",
-  },
-  priceInput: {
-    width: 300,
-    fontSize: 24,
-    borderWidth: 1,
-    borderRadius: 5,
-    padding: 10,
-  },
-  categoryScrollView: {
-    flex: 0.2,
-  },
-  memoInput: {
-    fontSize: 20,
-    flex: 0.4,
-    padding: 20,
-    margin: 20,
-    borderWidth: 1,
-    borderRadius: 5
-  },
-  buttonContainer: {
-    flex: 0.1,
-    alignItems: "center",
-    justifyContent: "center"
-  },
-  won: {
-    fontSize: 24,
-    margin: 10,
-    fontWeight: "bold"
-  }
-})
 
 
 export default function AddIncome() {
-  const [modalVisible, setModalVisible] = useState<boolean>(false);
-
-  const dispatch = useDispatch<AppDispatch>();
-  const userInput = useSelector((state: RootState) => {
-    return state.userInput;
-  })
-
-  function changePrice(v: string) {
-    try {
-      dispatch(userInputSlice.actions.update({price: parseInt(v)}))
-    } catch (error) {
-      console.log(error)
+  
+  const userInput = useSelector((state: RootState ) => state.userInput);
+  const iconSize = 50;
+  const icons = [
+    {
+      id: 1,
+      icon: <FontAwesome5 name="coins" size={iconSize} color={userInput.category == 1 ? "white" : "black"} />,
+      label: "월급"
+    },
+    {
+      id: 2,
+      icon: <FontAwesome5 name="money-bill-wave" size={iconSize} color={userInput.category == 2 ? "white" : "black"} />,
+      label: "용돈"
+    },
+    {
+      id: 3,
+      icon: <MaterialIcons name="attach-money" size={iconSize} color={userInput.category == 3 ? "white" : "black"} />,
+      label: "보너스"
+    },
+    {
+      id: 4,
+      icon: <AntDesign name="question" size={iconSize} color={userInput.category == 4 ? "white" : "black"} />,
+      label: "기타"
     }
-  } 
-  function onClickConfirm() {
-    try {
-      const { price, category, memo} = userInput;
-      const now = Date.now()
-
-      if(price <= 0) {
-        return;
-      }
-      const dataString = JSON.stringify({
-        type: "AddIncome",
-        price,
-        category,
-        memo,
-        date: now
-      });
-      AsyncStorage.setItem("transaction-history", dataString);
-      dispatch(userInputSlice.actions.clear());
-    } catch (error) {
-      console.log(error)
-    } finally {
-      router.back();
-    }
-  }
+  ]
 
   return(
-    <KeyboardAvoidingView style={styles.container}>
-      <View style={styles.priceInputContainer}>
-        <TextInput placeholder="1000" style={styles.priceInput} keyboardType="number-pad" onChangeText={changePrice}/>
-        <Text style={styles.won}>원</Text>
-      </View>
-      <View style={styles.categoryScrollView}>
-        <ScrollView horizontal={true}>
-          <AddIncomeList iconSize={50} />
-        </ScrollView>
-      </View>
-      <View style={styles.buttonContainer}>
-        <Button width={300} height={40} backgroundColor="black" text="메모 입력하기" fontColor="white" onPress={() => setModalVisible(!modalVisible)}/>
-        <MemoModal modalVisible={modalVisible} setModalVisible={setModalVisible} />
-      </View>
-      <View style={styles.buttonContainer}>
-        <Button width={150} height={40} backgroundColor="black" text="확인" fontColor="white" onPress={onClickConfirm}/>
-      </View>
-    </KeyboardAvoidingView>
+    <AddHistory icons={icons} type="income" />
   )
 }
