+++ app/Subtract.tsx
@@ -1,67 +1,54 @@
-import React, { useState } from "react";
-import { KeyboardAvoidingView, ScrollView, StyleSheet, TextInput, View } from "react-native";
-import SubtractList from "../components/Subtract/list";
-import Button from "../components/Button";
-import MemoModal from "../components/memoModal";
+import React from "react";
+import AddHistory from "../components/History/AddHistory";
+import { FontAwesome5, MaterialIcons, AntDesign, MaterialCommunityIcons } from '@expo/vector-icons';
+import { useSelector } from "react-redux";
+import { RootState } from "../redux/store";
 
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignContent: "center",
-  },
-  priceInputContainer: {
-    width: "100%",
-    justifyContent: "center",
-    flex: 0.15,
-    alignItems: "center",
-  },
-  priceInput: {
-    width: 300,
-    fontSize: 24,
-    borderWidth: 1,
-    borderRadius: 5,
-    padding: 10,
-  },
-  categoryScrollView: {
-    flex: 0.2,
-  },
-  memoInput: {
-    fontSize: 20,
-    flex: 0.4,
-    padding: 20,
-    margin: 20,
-    borderWidth: 1,
-    borderRadius: 5
-  },
-  buttonContainer: {
-    flex: 0.1,
-    alignItems: "center",
-    justifyContent: "center"
-  }
-})
 
 
 export default function AddIncome() {
-  const [modalVisible, setModalVisible] = useState<boolean>(false);
+  
+  const userInput = useSelector((state: RootState ) => state.userInput);
+  const iconSize = 50;
+  const icons = [ 
+    {
+      id: 1,
+      icon: <MaterialCommunityIcons name="silverware-fork-knife" size={iconSize} color={userInput.category == 1 ? "white" : "black"} />,
+      label: "식비"
+    },
+    {
+      id: 2,
+      icon: <FontAwesome5 name="briefcase-medical" size={iconSize} color={userInput.category == 2 ? "white" : "black"} />,
+      label: "의료"
+    },
+    {
+      id: 3,
+      icon: <FontAwesome5 name="bus" size={iconSize} color={userInput.category == 3 ? "white" : "black"} />,
+      label: "교통"
+    },
+    {
+      id: 4,
+      icon: <FontAwesome5 name="pen" size={iconSize} color={userInput.category == 4 ? "white" : "black"} />,
+      label: "자기계발"
+    },
+    {
+      id: 5,
+      icon: <FontAwesome5 name="shopping-bag" size={iconSize} color={userInput.category == 5 ? "white" : "black"} />,
+      label: "생필품"
+    },
+    {
+      id: 6,
+      icon: <MaterialIcons name="attach-money" size={iconSize} color={userInput.category == 6 ? "white" : "black"} />,
+      label: "저축"
+    },
+    {
+      id: 7,
+      icon: <AntDesign name="question" size={iconSize} color={userInput.category == 7 ? "white" : "black"} />,
+      label: "기타"
+    }, 
+  ]
 
   return(
-    <KeyboardAvoidingView style={styles.container}>
-      <View style={styles.priceInputContainer}>
-        <TextInput placeholder="1000원" style={styles.priceInput} keyboardType="number-pad" />
-      </View>
-      <View style={styles.categoryScrollView}>
-        <ScrollView horizontal={true}>
-          <SubtractList iconSize={50} />
-        </ScrollView>
-      </View>
-      <View style={styles.buttonContainer}>
-        <Button width={300} height={40} backgroundColor="black" text="메모 입력하기" fontColor="white" onPress={() => setModalVisible(!modalVisible)}/>
-        <MemoModal modalVisible={modalVisible} setModalVisible={setModalVisible} />
-      </View>
-      <View style={styles.buttonContainer}>
-        <Button width={150} height={40} backgroundColor="black" text="확인" fontColor="white" onPress={() => console.log('confirm')}/>
-      </View>
-    </KeyboardAvoidingView>
+    <AddHistory icons={icons} type="Spending" />
   )
 }
