+++ components/History/CategoryLIst.tsx
@@ -1,11 +1,10 @@
-import React, { useEffect } from "react";
-import { FontAwesome5, AntDesign, MaterialIcons } from '@expo/vector-icons';
-import { StyleSheet, Text, View } from "react-native";
-import { TouchableOpacity } from "react-native-gesture-handler";
+import React from "react";
 import { useSelector } from "react-redux";
 import { AppDispatch, RootState } from "../../redux/store";
+import { TouchableOpacity } from "react-native-gesture-handler";
+import { StyleSheet, Text, View } from "react-native";
+import { userInputSlice } from "../../redux/slice/userInput";
 import { useDispatch } from "react-redux";
-import { update, userInputSlice } from "../../redux/slice/userInput";
 
 
 
@@ -34,8 +33,12 @@ const styles = StyleSheet.create({
 })
 
 
-interface AddIncomeListProp {
-  iconSize: number,
+interface CategoryListProp {
+  icons: {
+    id: number,
+    icon: JSX.Element,
+    label: string
+  }[],
 }
 
 interface ItemProp {
@@ -45,6 +48,7 @@ interface ItemProp {
   id: number
 }
 
+
 function Item({icon, label, selected, id}: ItemProp) {
 
   const dispatch = useDispatch<AppDispatch>()
@@ -53,7 +57,7 @@ function Item({icon, label, selected, id}: ItemProp) {
     dispatch(userInputSlice.actions.update({category: id}))
   }
 
-
+  
   return(
     <TouchableOpacity style={[styles.container, selected ? styles.selectedContainer : null]} onPress={onPress}>
       <View style={[styles.icon]}>
@@ -65,39 +69,14 @@ function Item({icon, label, selected, id}: ItemProp) {
 
 }
 
-export default function AddIncomeList({iconSize}: AddIncomeListProp) {
-
+export default function CategoryList({ icons }: CategoryListProp) {
   const userInput = useSelector((state: RootState) => {
     return state.userInput
   });
-
-  const icons = [
-    {
-      id: 1,
-      icon: <FontAwesome5 name="coins" size={iconSize} color={userInput.category == 1 ? "white" : "black"} />,
-      label: "월급"
-    },
-    {
-      id: 2,
-      icon: <FontAwesome5 name="money-bill-wave" size={iconSize} color={userInput.category == 2 ? "white" : "black"} />,
-      label: "용돈"
-    },
-    {
-      id: 3,
-      icon: <MaterialIcons name="attach-money" size={iconSize} color={userInput.category == 3 ? "white" : "black"} />,
-      label: "보너스"
-    },
-    {
-      id: 4,
-      icon: <AntDesign name="question" size={iconSize} color={userInput.category == 4 ? "white" : "black"} />,
-      label: "기타"
-    }
-  ]
-
 
   return(
     <>
       {icons.map(v => <Item icon={v.icon} label={v.label} key={v.id} selected={v.id == userInput.category} id={v.id}/>)}
     </>
-  )
+  );
 }
